add back rss + sitemap + redirect rendering
