+++ Sources/ToucanSDK/PageBundle/PageBundle.swift
@@ -57,45 +57,3 @@ struct PageBundle {
             .sanitized()
     }
 }
-
-extension PageBundle {
-
-    /// Returns the context aware identifier, the last component of the slug
-    ///
-    /// Can be used when referencing contents, e.g.
-    /// slug: docs/installation
-    /// type: category
-    /// contextAwareIdentifier: installation
-    /// This way content can be identified, when knowing the type & id
-    var contextAwareIdentifier: String {
-        .init(slug.split(separator: "/").last ?? "")
-    }
-
-    func referenceIdentifiers(
-        for key: String
-    ) -> [String] {
-        var refIds: [String] = []
-        if let ref = frontMatter[key] as? String {
-            refIds.append(ref)
-        }
-        refIds += frontMatter[key] as? [String] ?? []
-        return refIds
-    }
-
-    func referenceIdentifiers(
-        for key: String,
-        join: ContentType.Join
-    ) -> [String] {
-        var refIds: [String] = []
-        switch join {
-        case .one:
-            if let ref = frontMatter[key] as? String {
-                refIds.append(ref)
-            }
-        case .many:
-            refIds = frontMatter[key] as? [String] ?? []
-        }
-        return refIds
-    }
-
-}
